feat: app.py 분리 + 구조화 로깅 + Alembic + PostgreSQL 지원 (fixes #11, fixes #12)
- app.py 1169줄 → routes/public.py + admin.py + api.py 분리 (app.py ~110줄)
- dictConfig 기반 JSON 구조화 로깅 (timestamp, level, logger, message)
- 보안 로그 별도 핸들러 (security logger → stderr)
- Alembic 마이그레이션 프레임워크 초기 설정
- PostgreSQL 연결 풀링 지원 (pool_size=10, max_overflow=20, pool_pre_ping)
- psycopg2-binary 의존성 추가
- Deduplicator 캐시 LRU 제한 (max 10000건, 초과 시 50% eviction)
