+++ src/web/routes/__init__.py
@@ -1,0 +1,5 @@
+from .public import router as public_router
+from .admin import router as admin_router
+from .api import router as api_router
+
+__all__ = ["public_router", "admin_router", "api_router"]
